payment_app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
PyMySQL==1.1.1
pydantic==2.11.7
python-dotenv==1.1.0
razorpay==1.4.2
setuptools==81.0.0
cryptography==42.0.0
PyJWT==2.8.1

